+++ servers/quarkus-server/src/test/java/org/projectnessie/server/TestCacheInvalidation.java
@@ -22,9 +22,9 @@ import static org.assertj.core.api.InstanceOfAssertFactories.list;
 import static org.projectnessie.quarkus.config.QuarkusStoreConfig.CONFIG_CACHE_INVALIDATIONS_SERVICE_NAMES;
 import static org.projectnessie.quarkus.config.QuarkusStoreConfig.CONFIG_CACHE_INVALIDATIONS_VALID_TOKENS;
 import static org.projectnessie.quarkus.config.QuarkusStoreConfig.NESSIE_VERSION_STORE_PERSIST;
-import static org.projectnessie.quarkus.providers.storage.CacheInvalidationReceiver.NESSIE_CACHE_INVALIDATION_TOKEN_HEADER;
-import static org.projectnessie.quarkus.providers.storage.CacheInvalidations.CacheInvalidationEvictObj.cacheInvalidationEvictObj;
-import static org.projectnessie.quarkus.providers.storage.CacheInvalidations.cacheInvalidations;
+import static org.projectnessie.server.distcache.CacheInvalidationReceiver.NESSIE_CACHE_INVALIDATION_TOKEN_HEADER;
+import static org.projectnessie.server.distcache.CacheInvalidations.CacheInvalidationEvictObj.cacheInvalidationEvictObj;
+import static org.projectnessie.server.distcache.CacheInvalidations.cacheInvalidations;
 import static org.projectnessie.versioned.storage.common.objtypes.StringObj.stringData;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -53,10 +53,10 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.projectnessie.nessie.relocated.protobuf.ByteString;
 import org.projectnessie.quarkus.config.QuarkusStoreConfig;
 import org.projectnessie.quarkus.providers.ServerInstanceId;
-import org.projectnessie.quarkus.providers.storage.CacheInvalidations;
-import org.projectnessie.quarkus.providers.storage.CacheInvalidations.CacheInvalidation;
-import org.projectnessie.quarkus.providers.storage.CacheInvalidations.CacheInvalidationEvictReference;
 import org.projectnessie.quarkus.util.HttpTestServer;
+import org.projectnessie.server.distcache.CacheInvalidations;
+import org.projectnessie.server.distcache.CacheInvalidations.CacheInvalidation;
+import org.projectnessie.server.distcache.CacheInvalidations.CacheInvalidationEvictReference;
 import org.projectnessie.versioned.storage.common.objtypes.Compression;
 import org.projectnessie.versioned.storage.common.objtypes.StringObj;
 import org.projectnessie.versioned.storage.common.persist.ObjId;
